// src/interfaces/disciplines.ts
export type DisciplineLevel = 1 | 2 | 3 | 4;

export enum DisciplineTree {
    CRUOR = "Cruor",
    ASSEMBLAGE = "Assemblage",
    SANCTUS = "Sanctus",
    MARTIAL = "Martial",
    SUBTERFUGE = "Subterfuge",
}

export enum DisciplineName {
    FOREBODE = "Forbode",
    FAMILIAR = "Familiar",
    LAST_LAUGH = "Last Laugh",
    BANISHED_KNOWLEDGE = "Banished Knowledge",
    SANCTUARY = "Sanctuary",
    EUTHANASIA = "Euthanasia",
    PERFECT_LOVE = "Perfect Love",
    IMPOSED_PARADIGM = "Imposed Paradigm",
    BFF = "BFF",
    SUMMONING = "Summoning",
    FORTUNE_TELLING = "Fortune Telling",
    RUINATION = "Ruination",
    POSSESSION = "Possession",
    HELPING_HANDS = "Helping Hands",
    THE_PESSIMIST = "The Pessimist",
    EXALTED_SUMMONING = "Exalted Summoning",
    CONDUIT_SUMMONING = "Conduit Summoning",
    THE_TOWER_REVERSED = "The Tower Reversed",
    FELLED_AFTERLIFE = "Felled Afterlife",
    FRIENDS_WITH_BENEFITS = "Friends With Benefits",
    RESONANCE = "Resonance",
    SOUL_BIND = "Soul Bind",
    IMBUED_FEAR = "Imbued Fear",
    CRUMBLING_TIME = "Crumbling Time",
    LIFE_TITHE = "Life Tithe",
    TO_FAUST = "To Faust",
    CORROSIVE_IDEAL = "Corrosive Ideal",
    OXYGEN_SIPHON = "Oxygen Siphon",
    GORE_SHOT = "Gore Shot",
    BRING_THE_SHADOWS = "Bring The Shadows",
    IN_DREAMS = "In Dreams",
    ROOT_THE_MIND = "Root The Mind",
    BLOOD_TOLL = "Blood Toll",
    BLOOD_PACT = "Blood Pact",
    IN_SECRET = "In Secret",
    WRETCHED_TETHER = "Wretched Tether",
    ORGANIC_LIQUIFY = "Organic Liquify",
    SOUL_STEAL = "Soul Steal",
    HE_WATCHES = "He Watches",
    SOUL_PACT = "Soul Pact",
    SINFUL_PRIVILEGE = "Sinful Privilege",
    LIFE_GRASP = "Life Grasp",
    ABADDONS_SUBTLETY = "Abaddon's Subtlety",
    INEPTITUDE = "Ineptitude",
    BLADE_WORKS = "Blade Works",
    ANTICIPATED_ATTACK = "Anticipated Attack",
    FORTUITOUS_HOMICIDE = "Fortuitous Homicide",
    HAMMER_HELM = "Hammer Helm",
    CRITICAL_DEFENSE = "Critical Defense",
    CHAINS_OF_COMMAND = "Chains Of Command",
    LIVING_BULWARK = "Living Bulwark",
    MAGE_BREAKER = "Mage Breaker",
    GIGAS_GRIP = "Giga's Grip",
    ONSLAUGHT = "Onslaught",
    BLOOD_RAGE = "Blood Rage",
    MORBID_ENVY = "Morbid Envy",
    WHIRLWIND = "Whirlwind",
    PROMISE = "Promise",
    PHYSICALLY_PERFECTED = "Physically Perfected",
    RELENTLESS_FEIGN = "Relentless Feign",
    DEAD_END = "Dead End",
    NEMESIS_ENFORCEMENT = "Nemesis Enforcement",
    BRING_THE_WRATH = "Bring The Wrath",
    WILL_TO_LIVE = "Will To Live",
    MASTER_OF_THE_VESSEL = "Master Of The Vessel",
    CHALLENGE_SUPREME = "Challenge Supreme",
    ASPECT = "Aspect",
    MAGIC_ARMOR = "Magic Armor",
    NIRVANAS_MARK = "Nirvana's Mark",
    INTERVENTION = "Intervention",
    GUARDIAN_ANGEL = "Guardian Angel",
    MEND = "Mend",
    COURAGE_STIFLE = "Courage Stifle",
    FROM_THE_GRAVE = "From The Grave",
    SOUL_CLEAR = "Soul Clear",
    ASCENSION = "Ascension",
    SALVATION = "Salvation",
    FERVOR = "Fervor",
    PROOF_OF_DIVINITY = "Proof Of Divinity",
    PARADIGM_ASSAULT = "Paradigm Assault",
    INVINCIBILITY = "Invincibility",
    WINGS_MANIFEST = "Wings Manifest",
    FONT_OF_RENEWAL = "Font Of Renewal",
    GREATER_EXISTENCE = "Greater Existence",
    THE_ONLY_TRUTH = "The Only Truth",
    DIMENSION_SHIFT = "Dimension Shift",
    INSPIRATION = "Inspiration",
    ZEALOUS_REPRISAL = "Zealous Reprisal",
    LIKE_THE_SHADOWS = "Like The Shadows",
    KILL_THE_MESSENGER = "Kill The Messenger",
    FOLLOW_THROUGH = "Follow Through",
    TRICK_SHOT = "Trick Shot",
    SPEED_INJUSTICE = "Speed Injustice",
    QUICK_BLOW = "Quick Blow",
    MURDER_CIRCUS = "Murder Circus",
    FALSE_IDOL = "False Idol",
    BLADE_DISTORTION = "Blade Distortion",
    DOWN_THE_MASSES = "Down The Masses",
    FRIEND_OVER_FOE = "Friend Over Foe",
    OVERWATCH = "Overwatch",
    SHAME_OF_THE_OTAKU = "Shame Of The Otaku",
    RAPID_RELOAD = "Rapid Reload",
    PROFOUND_INTUITION = "Profound Intuition",
    PRECISE_STRIKE = "Precise Strike",
    CQC = "C.Q.C",
    BRUTAL_SLAUGHTER = "Brutal Slaughter",
    FIST_OF_POLARIS = "Fist Of Polaris",
    NINJUTSU = "Ninjutsu",
    JACK_OF_ALL_BLADES = "Jack Of All Blades",
    BLOT_OUT_THE_SUN = "Blot Out The Sun",
}

export interface Discipline {
    level: DisciplineLevel;
    tree: DisciplineTree;
}

export type DisciplineData = {
    [key in DisciplineName]: Discipline;
};

export interface DisciplineTreeLevels {
    [key: string]: number;
}